egl: set TSD as NULL after deinit
When eglReleaseThread() is called from application's destructor (API with
__attribute__((destructor))), it crashes due to invalid memory access.

In this case, _egl_TLS is freed in the flow of _eglAtExit() as below but
_egl_TLS is not set to NULL.
    _eglDestroyThreadInfo
        _eglFiniTSD
            _eglAtExit
                __run_exit_handlers
                    exit

Later when the eglReleaseThread is called from application's destructor,
it ends-up accessing the freed _egl_TLS pointer.
    eglReleaseThread -> in libEGL_mesa
        eglReleaseThread -> in libEGL(glvnd)
            destructor() -> App's destructor

To resolve the invalid access, setting the _egl_TLS pointer as NULL
after freeing it.